test_hint_syntax now for both forms of prefab
